requirements/pip.txt
# region Base packages

pip==18.*

# endregion

# region Django

django
djangorestframework

# endregion
